error handler + sentry setup

+++ service/shopping-service.js
@@ -1,5 +1,4 @@
 const { shoppingrepository } = require('../Database');
-const { formatedata } = require('../Database/side-function/side1');
 
 // All Business logic will be here
 class shoppingservice {
@@ -9,29 +8,23 @@ class shoppingservice {
 
   async placeorder(userinput) {
     const { _id} = userinput;
-    try {
-      const orderresult = await this.repository.createneworder(_id);
-      return formatedata(orderresult);
-    } catch (err) {
-      throw err;
-    }
+    const existinguser = await this.repository.findusersbyid({ id });
+    if (!existinguser) throw new notfoundError("user not found by provided id")
+    return await this.repository.createneworder(_id);
+    
   }
 
   async deleteorders(userid,orderid) {
-    try {
-      const orders = await this.repository.deleteorder(userid,orderid);
-      return formatedata(orders);
-    } catch (err) {
-      throw err;
-    }
+    const existinguser = await this.repository.findusersbyid({ id });
+    if (!existinguser) throw new notfoundError("user not found by provided id")
+    return await this.repository.deleteorder(userid,orderid);
+    
   }
   async getorders(userid) {
-    try {
-      const orders = await this.repository.getorder(userid);
-      return formatedata(orders);
-    } catch (err) {
-      throw err;
-    }
+    const existinguser = await this.repository.findusersbyid({ id });
+    if (!existinguser) throw new notfoundError("user not found by provided id")
+    return await this.repository.getorder(userid);
+    
   }
 }
